Add screenrc config for amd docker use

+++ manifests/users/screen.pp
@@ -6,10 +6,10 @@
 # @summary A short summary of the purpose of this class
 #
 # NOTE - hybrid amd/nv not supported by the YET
+# NOTE - need to consolidate and/or refactor screenrc templates
 # @example
 #   include minebox::users::screen
 class minebox::users::screen {
-
 
   if $::minebox::nv_conf['enable'] == true {
     file { "/home/${minebox::miner_user}/.screenrc" :
@@ -26,18 +26,15 @@ class minebox::users::screen {
     }
   }
 
-  $mining_script = $::minebox::amd_conf['use_docker'] ? {
-    true  => '# nothing goes here for now, see module manifest:  minebox::users::screen',
-    false => "screen -t miner ${minebox::base_path}/scripts/claymore.sh",
-  }
+  $cron_screen = 'absent'
 
-  if $::minebox::amd_conf['enable'] == true {
+  if $::minebox::amd_conf['use_docker'] == true {
     file { "/home/${minebox::miner_user}/.screenrc" :
       ensure  => file,
       content => epp(
-        'minebox/amd/screenrc.epp',
+        'minebox/docker/screenrc.epp',
         {
-          'mining_script' => $mining_script,
+          'gpu_cfg' => $minebox::amd_conf['gpus'],
           }
         ),
       owner   => $minebox::miner_user,
@@ -45,17 +42,28 @@ class minebox::users::screen {
       mode    => '0774',
     }
   }
+  elsif $::minebox::amd_conf['enable'] == true {
+    $cron_screen = 'present'
 
-  $cron_screen = $minebox::amd_conf['use_docker'] ? {
-    true  => 'absent',
-    false => 'present',
-  }
+      file { "/home/${minebox::miner_user}/.screenrc" :
+        ensure  => file,
+        content => epp(
+          'minebox/amd/screenrc.epp',
+          {
+            'mining_script' => "screen -t miner ${minebox::base_path}/scripts/claymore.sh",
+            }
+          ),
+        owner   => $minebox::miner_user,
+        group   => $minebox::miner_group,
+        mode    => '0774',
+      }
 
-  cron { 'Screen Setup' :
-    ensure  => $cron_screen,
-    command => 'sleep 40 && /usr/bin/screen -d -m',
-    user    => $minebox::miner_user,
-    special => 'reboot',
+      cron { 'Screen Setup' :
+        ensure  => $cron_screen,
+        command => 'sleep 40 && /usr/bin/screen -d -m',
+        user    => $minebox::miner_user,
+        special => 'reboot',
+      }
   }
 
 }
